+++ sdk/src/test/java/com/hedera/hashgraph/sdk/FeeSchedulesTest.java
@@ -73,8 +73,9 @@ public class FeeSchedulesTest {
         var originalFeeSchedules = spawnFeeSchedulesExample();
         byte[] feeSchedulesBytes = originalFeeSchedules.toBytes();
         var copyFeeSchedules = FeeSchedules.fromBytes(feeSchedulesBytes);
-        assertThat(originalFeeSchedules.toString().equals(copyFeeSchedules.toString())).isTrue();
-        SnapshotMatcher.expect(originalFeeSchedules.toString()).toMatchSnapshot();
+        assertThat(originalFeeSchedules.toString().replaceAll("@[A-Za-z0-9]+", ""))
+            .isEqualTo(copyFeeSchedules.toString().replaceAll("@[A-Za-z0-9]+", ""));
+        SnapshotMatcher.expect(originalFeeSchedules.toString().replaceAll("@[A-Za-z0-9]+", "")).toMatchSnapshot();
     }
 
     @Test